assets/requirements.txt
pyyaml
requests
matplotlib
beautifulsoup4
pandas